+++ user_function/src/Encoder_MultiturnMag.c
@@ -1072,7 +1072,7 @@ uint8_t MulMag_Modbus_IsMyAddr(uint16_t addr)
 * 输出参量：对应地址的数据值
 * 编写日期：2026-2-24
 ****************************************************************************************/
-uint16_t MulMag_Modbus_Read(uint16_t addr)
+uint32_t MulMag_Modbus_Read(uint16_t addr)
 {
     switch (addr) {
         case 0x0200: return g_MulMag.Hall.Result;
